refactor(fw): Phase 4 LOG module — extract vayu_log to logger/, R1.2 -Werror
Sixth Phase-4 module (6/7).

R2.6 — text logging moved out of the utils/ catch-all: vayu_log +
vayu_log_queue extracted from src/utils/utils.c into new
src/logger/log_text.c. (utils.c keeps timestamp/device/CRC for now; those
go to SYS next.)

R2.1 — the logging API (vayu_log, vayu_log_queue, VAYU_LOG_QUEUE_SIZE) now
declared in the LOG umbrella include/logger/logger.h (still mirrored in
utils/utils.h transitionally until the SYS pass deletes that header).

R1.2 / CONV-06 — logger.c + log_text.c flipped to full R1.2 + -Werror;
fixed logger.c sign-conversion (file_size-1 -> (long) cast) and added an
explicit (uint8_t) cast on the log length.

Verification: arm target links clean (6 modules now under -Werror);
negative control confirms -Werror bites log_text.c; host SITL + full
5-suite ctest green. warning-rollout.md: LOG ✅.

+++ src/logger/logger.c
@@ -78,7 +78,7 @@ static void ensure_file_size(vfs_fd_t fd, uint32_t file_size) {
 
   if (current_size < (int)file_size) {
 
-    vfs_lseek(fd, file_size - 1, VFS_SEEK_SET);
+    vfs_lseek(fd, (long)(file_size - 1), VFS_SEEK_SET);
 
     uint8_t dummy = 0;
     vfs_write(fd, &dummy, 1);
